move aws_client_mock to include
